requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.11
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.3
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
